Add accept column filter in Support Training. (#57)
* Add accept column filter in Support Training.

* Move support adding filters into SupportPolicy

* Move accept filter into support policy
